refactor(badge): migrate badge module to a modular structure and improve code organization
