--- Heart/Badges/Domain/Actions/DeleteBadge.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Heart\Badges\Domain\Actions;
-
-use Heart\Badges\Domain\Repositories\BadgeRepository;
-
-class DeleteBadge
-{
-    public function __construct(private readonly BadgeRepository $badgeRepository)
-    {
-    }
-
-    public function handle(string $badgeId): void
-    {
-        $this->badgeRepository->delete($badgeId);
-    }
-}
