+++ app-modules/badge/composer.json
@@ -1,0 +1,23 @@
+{
+	"name": "he4rt/badge",
+	"description": "",
+	"type": "library",
+	"version": "1.0",
+	"license": "proprietary",
+	"require": {},
+	"autoload": {
+		"psr-4": {
+			"He4rt\\Badge\\": "src/",
+			"He4rt\\Badge\\Tests\\": "tests/",
+			"He4rt\\Badge\\Database\\Factories\\": "database/factories/"
+		}
+	},
+	"minimum-stability": "stable",
+	"extra": {
+		"laravel": {
+			"providers": [
+				"He4rt\\Badge\\Providers\\BadgeServiceProvider"
+			]
+		}
+	}
+}
