+++ app-modules/badge/src/Http/Requests/CreateBadgeRequest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace He4rt\Badge\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class CreateBadgeRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'provider' => ['required', 'in:discord,twitch'],
+            'name' => ['required', 'string'],
+            'description' => ['required', 'string'],
+            'image_url' => ['required', 'url'],
+            'redeem_code' => ['required', 'string'],
+            'active' => ['required', 'boolean'],
+        ];
+    }
+}
